backend/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
Flask-Limiter==3.5.0
PyJWT==2.8.0
PyMySQL==1.1.0
DBUtils==3.0.3
pandas==2.1.0
scikit-learn==1.3.0
bcrypt==4.0.1
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.8.5
pytest==7.4.0
black==23.7.0
flake8==6.1.0
